Conflicts:
	projects/bit_trade/bittrade/bittrade-admin/src/main/java/com/bittrade/admin/service/impl/c2c/TAdvertInfoServiceImpl.java
	projects/bit_trade/bittrade/bittrade-admin/src/main/java/com/bittrade/admin/service/impl/c2c/TAdvertOrderServiceImpl.java
	projects/bit_trade/bittrade/bittrade-c2c/src/main/java/com/bittrade/c2c/service/ITAdvertInfoService.java
	projects/bit_trade/bittrade/bittrade-c2c/src/main/java/com/bittrade/c2c/service/impl/TAdvertInfoServiceImpl.java
	projects/bit_trade/bittrade/bittrade-pojo-model/src/main/java/com/bittrade/pojo/model/TAdvertInfo.java
	projects/bit_trade/bittrade/bittrade-pojo-model/src/main/java/com/bittrade/pojo/model/TAdvertOrder.java

+++ projects/bit_trade/bittrade/bittrade-pojo-model/src/main/java/com/bittrade/pojo/model/TAdvertOrder.java
@@ -196,7 +196,7 @@ public class TAdvertOrder extends BaseModel<TAdvertOrder> {
 	/**
 	 * 广告类别 1出售 2购买
 	 */
-	private Byte advertType;
+	private Integer advertType;
 	
 	/**
 	 * 广告留言
